requirements.txt
langchain==0.3.15
langchain-openai==0.3.1
langgraph==0.2.65
langsmith==0.2.11
langchain_community==0.3.15
openai==1.58.1
pandas==2.2.1
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
scipy==1.12.0
numpy==1.26.4